Remove existing archive to avoid an upload failure
Github API for uploading "assets" will fail with "Unprocessable Entity"
if we try to upload an archive that is named the same as one of the
already uploaded archive.

This causes issues when rerunning the script and we have two options to
solve this problem:

1. Skip uploading of "duplicate" archive.
2. Or delete the old one and upload again.

Option 2 is more convenient as all checksums would be printed in a
single place, after a successful run. Thus this PR implements that
option.